fix(root): break root<->lights cycle by moving LightsController to lights/entities
