+++ apps/aurora-core/src/modules/root/socket-connection-manager.ts
@@ -4,7 +4,8 @@ import AsyncLock from 'async-lock';
 import HandlerManager from './handler-manager';
 import { AuthUser } from '../auth';
 import dataSource from '../../database';
-import { Audio, LightsController, Screen } from './entities';
+import { Audio, Screen } from './entities';
+import { LightsController } from '../lights/entities';
 import BaseLightsHandler from '../handlers/base-lights-handler';
 import { LightsGroup } from '../lights/entities';
 import { SECURE_NAMESPACES, SocketioNamespaces } from '../../socketio-namespaces';
